Create Full
This project is based on an algorithm presented to me by a college student working on his thesis to present,
called a gematria. Inside of the input element, you would enter your sentence or characters from the hebrew
language. For this example, we'll use Genesis 1:1 from the Torah. Simply copy and paste the text below:

ְּרֵאשִׁ֖ית בָּרָ֣א אֱלֹהִ֑ים אֵ֥ת הַשָּׁמַ֖יִם וְאֵ֥ת הָאָֽרֶץ:

The characters are written in a binary system, starting from 1 to 9, then 10, 20, 30 through 90, then 100, 200, 300
and 400. Some characters do share a value as is how the characters work in hebrew language.

The first algorithm takes the value of each letter entered multiplied against each other. The product is then
multiplied by how many letters there are.

The second algorithm takes each word separately and adds the values, then the sum of each word is multiplied
against each other. The product is then multiplied by the number of words there are.

The third algorithm takes the product of the letters and divides it by the product of the words. In this case, the
outcome of the algorithm comes out to the first 5 characters of Pi: 3.1415.